requirements.txt
streamlit
pandas
nltk
python-docx
PyPDF2
